move get_slack_support_messages script to new folder (#2358)
