+++ .github/workflows/get-support-channel-messages.yml
@@ -9,9 +9,9 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
-          python-version: "3.10"
-      - run: python3 -m pip install -r scripts/python/requirements.txt
-      - run: python3 scripts/python/get_slack_support_messages.py
+          python-version: "3.11"
+      - run: python3 -m pip install -r python/requirements.txt
+      - run: python3 -m python.scripts.get_slack_support_messages
         env:
           SLACK_BOT_TOKEN: ${{ secrets.SLACK_BOT_TOKEN }}
       - name: Report failure to Slack
